requirements.txt
slackclient
kubernetes